commands: Separate commands.go into its own package
Package "microovn/ovn" is getting a bit bloated. This may cause
problems with circular dependencies in the future if other packages
that require only some functions, are required to import whole "ovn"
package.

commands.go is ideal candidate for the separation from "ovn" package
as it does not use any other non-exported symbols from "ovn" and it
is very likely to be used by other packages to execute arbitrary
commands.
